+++ etna/analysis/forecast/plots.py
@@ -25,6 +25,8 @@ from sklearn.metrics import r2_score
 from statsmodels.graphics.gofplots import qqplot
 from typing_extensions import Literal
 
+from etna.analysis.forecast.utils import _check_metrics_df_empty_segments
+from etna.analysis.forecast.utils import _check_metrics_df_same_folds_for_each_segment
 from etna.analysis.forecast.utils import _prepare_forecast_results
 from etna.analysis.forecast.utils import _select_prediction_intervals_names
 from etna.analysis.forecast.utils import _validate_intersecting_segments
@@ -687,6 +689,11 @@ def plot_metric_per_segment(
 ):
     """Plot barplot with per-segment metrics.
 
+    If for some segment all metric values are missing, it isn't plotted, and the warning is raised.
+
+    If some segments have different set of folds with non-missing metrics,
+    it can lead to incompatible values between folds. The warning is raised in such case.
+
     Parameters
     ----------
     metrics_df:
@@ -715,11 +722,21 @@ def plot_metric_per_segment(
         if ``metric_name`` isn't present in ``metrics_df``
     NotImplementedError:
         unknown ``per_fold_aggregation_mode`` is given
+
+    Warnings
+    --------
+    UserWarning:
+        There are segments without non-missing metric values.
+    UserWarning:
+        Some segments have different set of folds to be aggregated on due to missing values.
     """
     if barplot_params is None:
         barplot_params = {}
 
     aggregation_mode = PerFoldAggregation(per_fold_aggregation_mode)
+
+    _check_metrics_df_empty_segments(metrics_df=metrics_df, metric_name=metric_name)
+    _check_metrics_df_same_folds_for_each_segment(metrics_df=metrics_df, metric_name=metric_name)
 
     plt.figure(figsize=figsize)
 
@@ -728,10 +745,13 @@ def plot_metric_per_segment(
 
     if "fold_number" in metrics_df.columns:
         metrics_dict = (
-            metrics_df.groupby("segment").agg({metric_name: aggregation_mode.get_function()}).to_dict()[metric_name]
+            metrics_df.groupby("segment")
+            .agg({metric_name: aggregation_mode.get_function()})
+            .dropna()
+            .to_dict()[metric_name]
         )
     else:
-        metrics_dict = metrics_df["segment", metric_name].to_dict()[metric_name]
+        metrics_dict = metrics_df[["segment", metric_name]].set_index("segment").dropna().to_dict()[metric_name]
 
     segments = np.array(list(metrics_dict.keys()))
     values = np.array(list(metrics_dict.values()))