+++ tests/test_analysis/test_forecast/test_plots.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import pytest
 
+from etna.analysis import plot_metric_per_segment
 from etna.analysis import plot_residuals
 from etna.analysis.forecast.plots import _get_borders_comparator
 from etna.metrics import MAE
@@ -50,3 +51,64 @@ def test_compare_error(segments_df):
 def test_compare(segments_df, expected):
     comparator = _get_borders_comparator(segment_borders=segments_df)
     assert comparator(name_a="a", name_b="b") == expected
+
+
+@pytest.fixture
+def metrics_df_with_folds() -> pd.DataFrame:
+    df = pd.DataFrame(
+        {
+            "segment": ["segment_0"] * 3 + ["segment_1"] * 3 + ["segment_2"] * 3,
+            "MAE": [1.0, 2.0, 3.0, 2.0, 3.0, 4.0, 3.0, 4.0, 5.0],
+            "MSE": [None, 3.0, 4.0, 3.0, 4.0, 5.0, 5.0, 6.0, 7.0],
+            "MAPE": [None, None, None, 20.0, 30.0, 40.0, 30.0, 40.0, 50.0],
+            "SMAPE": [None, None, None, None, None, None, 50.0, 60.0, 70.0],
+            "RMSE": [None, None, None, None, None, None, None, None, None],
+            "fold_number": [0, 1, 2, 0, 1, 2, 0, 1, 2],
+        }
+    )
+    return df
+
+
+@pytest.fixture
+def metrics_df_no_folds(metrics_df_with_folds) -> pd.DataFrame:
+    df = metrics_df_with_folds
+    df = df.groupby("segment").mean(numeric_only=False).reset_index().drop("fold_number", axis=1)
+    return df
+
+
+@pytest.mark.parametrize(
+    "df_name, metric_name",
+    [
+        ("metrics_df_with_folds", "MAE"),
+        ("metrics_df_no_folds", "MSE"),
+        ("metrics_df_no_folds", "MSE"),
+    ],
+)
+def test_plot_metric_per_segment_ok(df_name, metric_name, request):
+    metrics_df = request.getfixturevalue(df_name)
+    plot_metric_per_segment(metrics_df=metrics_df, metric_name=metric_name)
+
+
+@pytest.mark.parametrize(
+    "df_name, metric_name",
+    [
+        ("metrics_df_with_folds", "MAPE"),
+        ("metrics_df_no_folds", "RMSE"),
+    ],
+)
+def test_plot_metric_per_segment_warning_empty_segments(df_name, metric_name, request):
+    metrics_df = request.getfixturevalue(df_name)
+    with pytest.warns(UserWarning, match="There are segments with all missing metric values"):
+        plot_metric_per_segment(metrics_df=metrics_df, metric_name=metric_name)
+
+
+@pytest.mark.parametrize(
+    "df_name, metric_name",
+    [
+        ("metrics_df_with_folds", "MSE"),
+    ],
+)
+def test_plot_metric_per_segment_warning_non_comparable_segments(df_name, metric_name, request):
+    metrics_df = request.getfixturevalue(df_name)
+    with pytest.warns(UserWarning, match="Some segments have different set of folds to be aggregated on"):
+        plot_metric_per_segment(metrics_df=metrics_df, metric_name=metric_name)